Make TodayViewModel error nullable.
When an error happens we now only update the ui if the error is not
null. This is to prevent getting into a non error state like from
chossing a different day, doing a config change, and still having the error appear. Now the error is cleared when a new apod is fetched.